refactor: make config.errors private
