+++ src/deephaven_mcp/config/_errors.py
